fix(cli): validate packet count input with strtol instead of atoi
- Replaced atoi() with strtol() + endptr check in the -c option
  handler, so non-numeric input (e.g. "-c abc") is now rejected with
  a clear error instead of atoi() silently treating it as 0
- Changed packets_to_read from int to long in file_info_t and the
  corresponding limit variable in read_packets(), matching strtol()'s
  return type
- Wrapped the case 'c' block in braces to properly scope the new
  endptr/count local variables within the switch statement

+++ src/file_info.h
@@ -4,7 +4,7 @@
 typedef struct
 {
     char *path;
-    int packets_to_read;
+    long packets_to_read;
 } file_info_t;
 
 #endif
