+++ src/packet_reader.c
@@ -24,7 +24,7 @@ void read_packets(FILE *fp, uint32_t data_link_type, uint32_t magic_number, file
     packet_context_t ctx;
 
     unsigned int packet_counter = 0;
-    int limit = file_info->packets_to_read;
+    long limit = file_info->packets_to_read;
 
     while ((limit < 0 || packet_counter < (unsigned int)limit) && fread(&packet_header, sizeof(pcap_packet_header_t), 1, fp) == 1)
     {
